Add Travis-CI integration, based on what I contributed to dbus
The scripts in tests/ are reasonably CI-system-agnostic and should
be feasible to hook up to Gitlab-CI or similar if desired (as has been
done in dbus). It should also be reasonably straightforward to build
in a non-Debian-derived container by adding an alternative code path
alongside the various apt-get calls.

If this is merged, contributors can enable CI for their forks even if
it isn't enabled for the upstream repository.

Signed-off-by: Simon McVittie <smcv@collabora.com>