Add support for pipelined SASL handshakes, such as sdbus and zbus
sd-bus sends the BEGIN command in at the same time as the AUTH command,
assuming that the authentication will succeed. This ends up confusing
xdg-dbus-proxy, as it assumes that BEGIN is not sent at the same time as
any other messages. As a result, the server's authentication
replies end up interpreted as standard D-Bus messages, failing parsing
and resulting in the client being unable to connect.

This changes the code to keep track of the number of authentication
requests and the number ofresponses from the server to know when the
authentication phase of the connection has actually completed.

Until the authentication is completed, hold off all client messages
received after BEGIN. We only send to the server the authentication
commands up (and including) the BEGIN command, but everything past the
line terminators is queued into the `extra_input_data` buffer and we
stop reading from the client socket.

Once the authentication is completed, the (partial) message we saved
into `extra_input_data` is queued for the D-Bus server into the first
outgoing buffer, and reading from the client socket resumes.

Note that in order to cleanly process the partial sendind of data,
another offset is introduced into the Buffer structure, which holds the
bytes of buffers which have already been sent over the socket.

Fixes #21 (finally!)

This is based on initial work by:
 * Ryan Gonzalez <ryan.gonzalez@collabora.com>
 * Alberto Mardegan <a.mardegan@omp.ru>