Add an option to disable build-time tests
When using xdg-dbus-proxy as a subproject in Flatpak, we don't
necessarily want to run its unit tests as part of Flatpak's test
suite.

In particular, when bundling selected source code from xdg-dbus-proxy
into Flatpak `make dist` tarballs, it would be simpler if we could
avoid having to include the unit tests.

Signed-off-by: Simon McVittie <smcv@collabora.com>