Query name owners just once
Since all clients connecting to the proxy share the same policy filters,
we can setup the name owner data just once, when the first client
connects, and store it in the FlatpakProxy object so that all other
clients can use it.

This avoid duplicating the calls to ListNames and GetNameOwner.

Fixes: #55